requirements.txt
flask
flask-cors
nanoid
requests
pymongo
cryptography
marshmallow
python-dotenv